Wed, Jun 1, 2022, 5:33 AM +00:00

+++ credit/test.c
@@ -4,7 +4,7 @@
 
 int main(void)
 {
-    int input = get_long("Number? ");
-    int mod = input % 10;
-    printf("Mod: %i\n", mod);
+    long input = get_long("Number? ");
+    long mod = input % 10;
+    printf("Mod: %li\n", mod);
 }
